refactor(evals): move red-team code from @kbn/evals to @kbn/evals-extensions
Moves the Phase 3 red-teaming attack modules, guardrails engine, and
runner from `@kbn/evals` to `@kbn/evals-extensions` to follow the
architecture where extensions live in a separate package that depends
on `@kbn/evals` (never the reverse).

Reverts evaluator modifications in `@kbn/evals` to match upstream/main
since those changes are now canonical.
