+++ x-pack/platform/packages/shared/kbn-evals-extensions/index.ts
@@ -14,69 +14,36 @@
  * ## Architecture
  *
  * Dependency flow:
- * - ✅ kbn-evals-extensions → imports from → kbn-evals
- * - ❌ kbn-evals → MUST NOT import from → kbn-evals-extensions
- *
- * Evaluation suites can opt-in to extensions by importing directly:
- *
- * @example
- * ```typescript
- * import { evaluate } from '@kbn/evals';
- * import { createToxicityEvaluator, costTracker } from '@kbn/evals-extensions';
- *
- * evaluate('test', async ({ executorClient }) => {
- *   await executorClient.runExperiment(
- *     { dataset, task },
- *     [createToxicityEvaluator()]  // Extension evaluator
- *   );
- *   await costTracker.logRunCost(runId);  // Extension feature
- * });
- * ```
- *
- * ## Roadmap
- *
- * Features are being added incrementally:
- * - **PR #1**: Foundation (current) - Package setup, no functional changes
- * - **PR #2**: Cost tracking & metadata
- * - **PR #3**: Dataset management utilities
- * - **PR #4**: Safety evaluators (toxicity, PII, bias, etc.)
- * - **PR #5**: UI components (run comparison, example explorer)
- * - **PR #6**: DX enhancements (watch mode, caching, parallel execution)
- * - **PR #7**: Advanced analytics (confidence intervals, outlier detection)
- * - **PR #8**: A/B testing & active learning
- * - **PR #9**: Human-in-the-loop workflows
- * - **PR #10**: IDE integration (VS Code extension, Cursor skills)
+ * - kbn-evals-extensions -> imports from -> kbn-evals
+ * - kbn-evals -> MUST NOT import from -> kbn-evals-extensions
  *
  * @packageDocumentation
  */
 
-// Re-export core types from kbn-evals for convenience
-// This allows users to import from one place, but doesn't create reverse dependency
-export type { Evaluator, Example, EvaluationDataset, TaskOutput } from '@kbn/evals';
+export {
+  createRedTeamRunner,
+  createGuardrailsEngine,
+  DEFAULT_GUARDRAIL_RULES,
+  promptInjectionModule,
+  privilegeEscalationModule,
+  infoExtractionModule,
+  jailbreakingModule,
+  ALL_MODULES,
+  MODULES_BY_NAME,
+} from './src/red_team';
 
-export type { EvaluationScoreDocument } from '@kbn/evals';
-
-/**
- * Extension-specific types (to be populated in future PRs)
- */
-export interface ExtensionConfig {
-  /**
-   * Configuration for extension features
-   * Will be expanded as features are added
-   */
-  placeholder?: string;
-}
-
-/**
- * Feature exports (to be populated in future PRs)
- *
- * Examples of what will be exported:
- * - export { createToxicityEvaluator } from './src/evaluators/safety/toxicity';
- * - export { costTracker } from './src/tracking/cost_calculator';
- * - export { watchMode } from './src/execution/watch_mode';
- * - export { createABTest } from './src/experimentation/ab_testing/framework';
- * - export { reviewQueue } from './src/human_review/workflow/review_workflow';
- */
-
-// Placeholder export to ensure package builds
-export const EVALS_EXTENSIONS_VERSION = '1.0.0';
+export type {
+  AttackCategory,
+  Severity,
+  AttackExample,
+  AttackModule,
+  AttackModuleConfig,
+  AttackResult,
+  RedTeamRunSummary,
+  GuardrailRule,
+  GuardrailAction,
+  GuardrailMatch,
+  GuardrailCheckResult,
+  RedTeamRunnerConfig,
+  TaskResult,
+} from './src/red_team';
