+++ x-pack/platform/packages/shared/kbn-evals-extensions/src/red_team/guardrails.test.ts
