+++ x-pack/platform/packages/shared/kbn-evals-extensions/src/red_team/modules/info_extraction.ts
