+++ x-pack/platform/packages/shared/kbn-evals-extensions/src/red_team/modules/prompt_injection.ts
