--- x-pack/platform/packages/shared/kbn-evals-extensions/src/utils/index.ts
@@ -1,19 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-/**
- * Utility functions for @kbn/evals-extensions
- *
- * Will be populated in future PRs with:
- * - Common helpers
- * - Shared calculations
- * - Type guards
- * - Validation utilities
- */
-
-// Placeholder export
-export const UTILS_VERSION = '1.0.0';
