+++ x-pack/platform/packages/shared/kbn-evals/src/cli/index.ts
@@ -17,7 +17,6 @@ import { startCmd } from './commands/start';
 import { stopCmd } from './commands/stop';
 import { logsCmd } from './commands/logs';
 import { scoutCmd } from './commands/scout';
-import { redTeamCmd } from './commands/red_team';
 
 export async function run() {
   await new RunWithCommands(
@@ -36,7 +35,6 @@ export async function run() {
       envCmd,
       ciMapCmd,
       compareCmd,
-      redTeamCmd,
     ]
   ).execute();
 }
