+++ x-pack/platform/packages/shared/kbn-evals/src/evaluators/similarity/index.ts
@@ -65,6 +65,15 @@ function sortKeys(value: unknown): unknown {
     }, {});
 }
 
+/**
+ * Computes term-frequency cosine similarity between expected and actual outputs.
+ *
+ * Both inputs are normalized to lowercase tokens. Objects are sorted by keys and
+ * serialized to JSON for consistent comparison. Returns a score between 0 and 1,
+ * with a configurable threshold for the similar/dissimilar label.
+ *
+ * @param config.threshold - Minimum cosine similarity to be labeled 'similar' (default: 0.7)
+ */
 export function createSimilarityEvaluator(config?: { threshold?: number }): Evaluator {
   const threshold = config?.threshold ?? 0.7;
