+++ x-pack/platform/packages/shared/kbn-evals/src/evaluators/trajectory/index.ts
@@ -40,6 +40,18 @@ function computeLCS(a: string[], b: string[]): string[] {
   return lcs;
 }
 
+/**
+ * Evaluates tool-call sequence alignment against a golden path using Longest Common
+ * Subsequence (LCS) for order scoring and set intersection for coverage scoring.
+ *
+ * The final score is a weighted combination of order and coverage scores.
+ * Both weights must sum to 1.
+ *
+ * @param config.extractToolCalls - Extracts actual tool call names from task output
+ * @param config.goldenPathExtractor - Extracts expected tool sequence from ground truth
+ * @param config.orderWeight - Weight for LCS-based order score (default: 0.5)
+ * @param config.coverageWeight - Weight for set-based coverage score (default: 0.5)
+ */
 export function createTrajectoryEvaluator(config: {
   extractToolCalls: (output: unknown) => string[];
   goldenPathExtractor: (expected: unknown) => string[];
